backend/shuscribe/pyproject.toml
[project]
name = "shuscribe"
version = "0.1.0"
description = "Narrative Wiki Generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "anthropic>=0.49.0",
    "asyncpg>=0.30.0",
    "celery>=5.4.0",
    "fastapi[standard]>=0.115.11",
    "google-genai>=1.7.0",
    "httpx>=0.28.1",
    "jsonref>=1.1.0",
    "openai>=1.66.5",
    "pgvector>=0.4.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.39",
    "sse-starlette>=2.2.1",
    "supabase>=2.13.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.1.2",
    "isort>=6.0.1",
    "jupyter>=1.1.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "ruamel-yaml>=0.18.10",
]
